requirements.txt
python-dotenv>=1.0.1
streamlit>=1.38.0
langchain>=0.2.16
langchain-openai>=0.1.22
pydantic>=2.8.2
SQLAlchemy>=2.0.30
tqdm>=4.66.4
requests>=2.31.0
